refactor: split util into smaller modules
